try to also fix the issue where proot might not match the filesystem architecture
