+++ app/build.gradle
@@ -198,7 +198,7 @@ ext.architectures = ["armeabi-v7a", "arm64-v8a", "x86", "x86_64"]
 ext.libDir = "$project.projectDir/src/main/jniLibs"
 
 task downloadAssets(type: Download) {
-    def assetVersion = "v1.0.3"
+    def assetVersion = "v1.0.5"
     def baseUrl = "https://github.com/CypherpunkArmory/UserLAnd-Assets-Support/releases/download/$assetVersion"
     for (arch in architectures) {
         src "$baseUrl/$arch-assets.zip"
@@ -222,6 +222,7 @@ task fetchAssets(dependsOn: downloadAssets) {
                 // Lib files must start with 'lib' and end with '.so.'
                 rename '(.*)', 'lib_$1.so'
             }
+            new File("$libDir/$arch","lib_arch.so").text = "$arch"
         }
     }
 }
