+++ app/src/main/java/tech/ula/model/remote/GithubApiClient.kt
@@ -7,9 +7,9 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import okhttp3.OkHttpClient
 import okhttp3.Request
-import tech.ula.utils.DeviceArchitecture
 import tech.ula.utils.Logger
 import tech.ula.utils.SentryLogger
+import tech.ula.utils.UlaFiles
 import java.io.IOException
 import java.net.UnknownHostException
 
@@ -20,9 +20,9 @@ class UrlProvider {
 }
 
 class GithubApiClient(
-    private val deviceArchitecture: DeviceArchitecture = DeviceArchitecture(),
-    private val urlProvider: UrlProvider = UrlProvider(),
-    private val logger: Logger = SentryLogger()
+        private val ulaFiles: UlaFiles,
+        private val urlProvider: UrlProvider = UrlProvider(),
+        private val logger: Logger = SentryLogger()
 ) {
     private val client = OkHttpClient()
     private val latestResults: HashMap<String, ReleasesResponse?> = hashMapOf()
@@ -39,7 +39,7 @@ class GithubApiClient(
     suspend fun getAssetsListDownloadUrl(repo: String): String = withContext(Dispatchers.IO) {
         val result = latestResults[repo] ?: queryLatestRelease(repo)
 
-        return@withContext result.assets.find { it.name == "${deviceArchitecture.getArchType()}-assets.txt" }!!.downloadUrl
+        return@withContext result.assets.find { it.name == "${ulaFiles.getArchType()}-assets.txt" }!!.downloadUrl
     }
 
     @Throws(IOException::class)
@@ -52,7 +52,7 @@ class GithubApiClient(
     @Throws(IOException::class)
     suspend fun getAssetEndpoint(assetType: String, repo: String): String = withContext(Dispatchers.IO) {
         val result = latestResults[repo] ?: queryLatestRelease(repo)
-        val assetName = "${deviceArchitecture.getArchType()}-$assetType"
+        val assetName = "${ulaFiles.getArchType()}-$assetType"
 
         return@withContext result.assets.find { it.name == assetName }!!.downloadUrl
     }
