+++ app/src/main/java/tech/ula/model/repositories/AssetRepository.kt
@@ -24,7 +24,8 @@ data class DownloadMetadata(
 class AssetRepository(
     private val applicationFilesDirPath: String,
     private val assetPreferences: AssetPreferences,
-    private val githubApiClient: GithubApiClient = GithubApiClient(),
+    private val ulaFiles: UlaFiles,
+    private val githubApiClient: GithubApiClient = GithubApiClient(ulaFiles),
     private val httpStream: HttpStream = HttpStream(),
     private val logger: Logger = SentryLogger()
 ) {
